Move db-export services to data folder
https://eaflood.atlassian.net/browse/WATER-4057

We are on the cusp of adding our second bill run type; two-part tariff. Our experience of implementing the first, supplementary billing has shown that we're going to need a lot of services! 😁

So, ahead of starting to implement we are doing some housekeeping in our services folder. The first of these is to move the db-export services into `app/services/data`. We also tweak the name slightly to just `/export` to be more inline with the others (`/seed` and `/tear-down`) i.e. named after what they do rather than what they are doing it to.

We already access this functionality from the `/data` endpoint anyway so this also resolves an inconsistency in the repo structure.
