+++ app/controllers/data/data.controller.js
@@ -7,15 +7,17 @@
 
 const Boom = require('@hapi/boom')
 
-const DbExportService = require('../../services/db-export/db-export.service.js')
+const ExportService = require('../../services/data/export/export.service.js')
 const SeedService = require('../../services/data/seed/seed.service.js')
 const TearDownService = require('../../services/data/tear-down/tear-down.service.js')
 
 /**
  * Triggers export of all relevant tables to CSV and then uploads them to S3
+ *
+ * > Has to be called something other than 'export' because export is a reserved word
  */
-async function dbExport (_request, h) {
-  DbExportService.go()
+async function exportDb (_request, h) {
+  ExportService.go()
 
   return h.response().code(204)
 }
@@ -41,7 +43,7 @@ async function tearDown (_request, h) {
 }
 
 module.exports = {
-  dbExport,
+  exportDb,
   seed,
   tearDown
 }
