+++ test/services/data/export/export-data-files.service.test.js
@@ -12,7 +12,7 @@ const fs = require('fs')
 const path = require('path')
 
 // Thing under test
-const ExportDataFilesService = require('../../../app/services/db-export/export-data-files.service.js')
+const ExportDataFilesService = require('../../../../app/services/data/export/export-data-files.service.js')
 
 const csvHeader = [
   '"billingChargeCategoryId"',
