+++ test/services/data/export/send-to-s3-bucket.service.test.js
@@ -12,7 +12,7 @@ const { expect } = Code
 const { S3Client, PutObjectCommand } = require('@aws-sdk/client-s3')
 
 // Thing under test
-const SendToS3BucketService = require('../../../app/services/db-export/send-to-s3-bucket.service.js')
+const SendToS3BucketService = require('../../../../app/services/data/export/send-to-s3-bucket.service.js')
 
 describe('Send to S3 bucket service', () => {
   let s3Stub
